Time: 25 ms (75.46%), Space: 47.6 MB (81.48%) - LeetHub

+++ 2222-number-of-ways-to-select-buildings/2222-number-of-ways-to-select-buildings.java
@@ -3,7 +3,7 @@ class Solution {
         // 101
         // 010
         long res = 0;
-        int[] cnt = new int[2];
+        long[] cnt = new long[2];
         for (int i = 0; i < s.length(); i++) {
             cnt[s.charAt(i) - '0']++;
         }
